Update Select.tsx
Suggestion of modifications:

Adding a sorting option:
A "Minted" and "Not Minted" option is added to the dropdown menu to allow filtering of the badges based on their status.

Managing minted badges:
A mintedItems array is passed as a prop, containing the keys of badges already minted by the user. The filteredItems function filters the badges to display only those that match the selected filter.

Filtering logic:
If the "Not Minted" option is selected, only badges that have not yet been minted (i.e., those not present in mintedItems) will be displayed. By default, all badges are shown.